app/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
pydantic==2.10.6
pydantic-settings==2.8.0
python-dotenv==1.0.1
supabase==2.13.0
langchain==0.3.19
langchain-community==0.3.18
langchain-openai==0.3.7
langchain-together==0.3.0
faiss-cpu==1.10.0
openai==1.64.0
together==1.4.1
